+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 13.4.0
 
 * Use the name of signon instead of signonotron2 since it was renamed.
 * Allow running a mock sso session in a Rails production environment via ENV